distribute/pyproject.toml
[project]
name = "event-data-pipeline"
version = "0.1.0"
description = "Event data pipeline using PySpark and Apache Airflow"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "pyspark>=4.1.1",
    "pyyaml>=6.0.3"
]
